# packages/shared/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: workspace-chat API
  version: 0.0.0
  description: >-
    エラーの応答の本体はすべて ErrorResponse である。仕様に無いパス（前置き /api の外を含む）は
    404（not_found）を返し、パスごとには載せない。
servers:
  # 全ルートの前置きは /api（#77 の決定。apps/api/src/app-setup.ts の createApp が付ける）。
  - url: /api
# 既定ですべての操作にアクセストークンを求める。認証を要さない操作だけが `security: []` で外す
# （api 側は auth/access-token.guard.ts が同じ既定を持つ。一致は auth/access-token.test.ts が仕様を読んで確かめる）。
security:
  - bearerAuth: []
paths:
...
  /workspaces/{id}/channels/{channelId}/pins:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/ChannelId'
    get:
      operationId: listPins
      summary: チャンネルのピン留めの一覧（F-33）
      description: >-
        ピン留めしたメッセージを、ピン留めした新しい順にすべて返す（1つのチャンネルに最大 100 件。機能一覧 13.2）。
        削除済みのメッセージは返さない。コードは2段階で決まる: 所属していなければ種別によらず 404。所属していて参加していなければ、
        パブリックは 403 not_a_channel_member・プライベートは 404。オーナーでも参加していなければ同じ（オーナーの例外はメッセージに及ばない。機能一覧 3.1・4.1）。
        アーカイブ済みでも参加者は読める（機能一覧 3.2）。ピン留めの変化は配信しない（要件定義書 4.1 の配信の対象イベントに無い）
      responses:
        '200':
          description: ピン留めしたメッセージ（ピン留めした新しい順）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{id}/channels/{channelId}/messages/{messageId}/pin:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/ChannelId'
      - $ref: '#/components/parameters/MessageId'
    put:
      operationId: pinMessage
      summary: メッセージをピン留めする（F-33）
      description: >-
        既にピン留めしていれば何も変えず、最初にピン留めした人と時刻を 200 で返す。スレッドの返信もピン留めできる。
        判定の順: 所属していなければ 404 → 参加していなければパブリックは 403 not_a_channel_member・プライベートは 404（オーナーでも同じ）→
        そのチャンネルに無い・削除済みのメッセージは 404 → アーカイブ済みのチャンネルは 409 channel_archived →
        そのチャンネルのピン留め（削除済みのメッセージを除く）が 100 件に達していれば 409 pin_limit_reached（実装時に決めた値）。
        配信しない（要件定義書 4.1 の配信の対象イベントに無い）。利用者ごとに1分 60 回まで（外すのとは別枠。実装時に決めた値）
      responses:
        '200':
          description: ピン留めしたメッセージ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinnedMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '409':
          description: アーカイブ済みのチャンネル（channel_archived）か、ピン留めの件数の上限（pin_limit_reached）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: unpinMessage
      summary: メッセージのピン留めを外す（F-33）
      description: >-
        そのチャンネルの参加者なら、ピン留めした本人でなくても外せる（付けられる人と外せる人を揃える。機能一覧 13.2 の「判断が必要な点」の決定）。
        ピン留めしていなければ何も変えずに 204 を返す。判定の順はピン留めと同じ（件数の上限を除く）。
        配信しない。利用者ごとに1分 60 回まで（ピン留めとは別枠。実装時に決めた値）
      responses:
        '204':
          description: 外した（ピン留めしていなかった場合も 204）
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: パブリックチャンネルに参加していない（not_a_channel_member）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '409':
          description: アーカイブ済みのチャンネル（channel_archived）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      description: ログイン（/auth/login）とリフレッシュ（/auth/refresh）が返すアクセストークン（JWT・HS256）
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      description: チャンネルの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
    MessageId:
      name: messageId
      in: path
      required: true
      description: メッセージの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
    WorkspaceId:
      name: id
      in: path
      required: true
      description: ワークスペースの id。形が uuid でなければ 400
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: >-
        Authorization ヘッダーが無い、または Bearer 方式でない（authentication_required）か、
        Bearer 方式で資格情報が無い・壊れている・期限切れ・利用者が退会済み（invalid_token。どれに当たったかは区別しない。機能一覧 1.4）。
        アクセストークンはリフレッシュ（/auth/refresh）で取り直す。仕様の形の検証（400）はトークンの確認より先に行う（仕様で書けない検証は後に行う）
      headers:
        WWW-Authenticate:
          description: >-
            Authorization ヘッダーが無い・Bearer 方式でないときは `Bearer`、Bearer 方式で資格情報が無い・使えないときは `Bearer error="invalid_token"`（RFC 6750 3.1）
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: >-
        入力が仕様に合わない（validation_failed）か、本体を JSON として読めない（invalid_body）。
        どちらも送られた値を応答に載せない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: 仕様に無いメソッド（method_not_allowed）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: レート制限を超えた（too_many_requests。上限はパスの説明に書く）
      headers:
        Retry-After:
          description: 再び受け付けるまでの秒数
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: >-
        見つからない（not_found）。所属していないワークスペースは、存在の有無を区別せずこれになる（機能一覧 2.1・1.4）。
        本体は状態コードの本体であり、権限の有無を述べない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 想定外の失敗（internal_error）。例外のメッセージは載せない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          description: >-
            エラーの種類。api が返す値はこの列挙だけであり、api と web は生成した型で同じ列挙を使う
            （綴りを誤ると型検査で落ちる）
          type: string
          enum:
            - validation_failed
            - invalid_body
            - not_found
            - method_not_allowed
            - payload_too_large
            - unsupported_media_type
            - too_many_requests
            - user_id_taken
            - registration_disabled
            - invalid_credentials
            - authentication_required
            - invalid_token
            - csrf_rejected
            - owner_only
            - invitee_not_found
            - already_invited
            - already_member
            - owner_cannot_leave
            - password_mismatch
            - owner_cannot_delete_account
            - pin_limit_reached
            - channel_name_taken
            - not_a_channel_member
            - already_channel_member
            - channel_archived
            - channel_not_private
            - channel_not_archived
            - not_message_author
            - dm_with_self
            - dm_counterpart_not_found
            - dm_counterpart_unavailable
            - reaction_limit_reached
            - unsupported_file_type
            - file_too_large
            - upload_not_received
            - upload_in_progress
            - attachment_unavailable
            - request_rejected
            - internal_error
        message:
          type: string
        errors:
          description: 入力の検証で落ちた箇所。送られた値は含めない
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - path
              - message
            properties:
              path:
                type: string
              message:
                type: string
...
    UserSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - displayName
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        displayName:
          type: string
...
    Attachment:
      description: 添付ファイル（機能一覧 11.1・11.2）
      type: object
      additionalProperties: false
      required:
        - id
        - fileName
        - contentType
        - kind
        - size
        - url
      properties:
        id:
          description: アップロードの識別子
          type: string
          format: uuid
        fileName:
          description: 利用者が付けた元のファイル名（表示用。保存名とは違う）
          type: string
        contentType:
          description: 配信の Content-Type（検証した形式からサーバーが決めた値。テキスト系は text/plain; charset=utf-8）
          type: string
        kind:
          description: 種別（画像は表示し、動画は #t=30 の位置を表示し、それ以外はリンクにする。機能一覧 11.1）
          type: string
          enum:
            - image
            - video
            - document
            - archive
        size:
          description: 大きさ（バイト）
          type: integer
          minimum: 1
        url:
          description: >-
            配信 URL のパス（/files/workspace/{workspaceId}/channel/{channelId}/{id}/{保存名}）。CloudFront の署名付き Cookie で取得する（機能一覧 11.2）
          type: string
...
    Message:
      description: チャンネルのメッセージ（機能一覧 4.1・4.2）
      type: object
      additionalProperties: false
      required:
        - id
        - channelId
        - author
        - body
        - createdAt
        - editedAt
        - deleted
        - parentId
        - replyCount
        - replyParticipants
        - mentions
        - reactions
        - attachments
      properties:
        id:
          description: UUIDv7。一覧のカーソル（before）にそのまま使う
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        parentId:
          description: スレッドの親の id。チャンネルの本体のメッセージなら null（機能一覧 6）
          type:
            - string
            - 'null'
          format: uuid
        replyCount:
          description: 削除されていない返信の件数。返信なら常に 0（返信はスレッドを持たない。機能一覧 6）
          type: integer
          minimum: 0
        replyParticipants:
          description: >-
            返信した参加者。削除されていない返信を書いた、退会していない利用者を、最後に返信した順に最大3人（機能一覧 6）。
            返信と、返信の無いメッセージは空
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/UserSummary'
        author:
          description: 投稿者。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        body:
          description: 本文。削除済みなら null（本文は DB に残るが返さない。機能一覧 4.2）
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
        editedAt:
          description: 最後に編集した時刻。編集していなければ null（「編集済み」の表示に使う。機能一覧 4.2）
          type:
            - string
            - 'null'
          format: date-time
        deleted:
          description: 削除済みか（「このメッセージは削除されました」に置き換えて表示する。機能一覧 4.2）
          type: boolean
        mentions:
          description: >-
            本文のメンションの対象（機能一覧 9.1）。投稿・返信のときに、そのチャンネルの参加者で退会していない利用者へ解決できた
            `@ユーザーID` を保存し、本文に最初に現れた順に1人1件で載せる。編集では、本文にユーザーID が残っている対象は解決し直さずに残す
            （チャンネルを抜けた・退会した対象も残り、退会した対象は user が null）。本文から消えた対象は外し、新しく書いた `@` は投稿と同じく解決して足す。
            削除済みのメッセージは空
          type: array
          items:
            $ref: '#/components/schemas/MessageMention'
        reactions:
          description: >-
            リアクション（機能一覧 7）。絵文字ごとに1件で、その絵文字を初めて付けた順。削除済みのメッセージは空
            （本文を返さないのと同じく返さない。付けた記録は DB に残る）
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/Reaction'
        attachments:
          description: >-
            投稿に付けた添付（上げた順）。削除済みのメッセージは空（本文を返さないのと同じく返さない。機能一覧 4.2）
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    Reaction:
      description: 1つの絵文字のリアクション（機能一覧 7）
      type: object
      additionalProperties: false
      required:
        - emoji
        - count
        - users
      properties:
        emoji:
          description: 絵文字1つ
          type: string
        count:
          description: >-
            付けた人数（カウンタ列。一覧で COUNT を発行しない。要件定義書 4.1）。退会した利用者の分も数える
          type: integer
          minimum: 1
        users:
          description: >-
            付けた人（「誰が付けたか」をホバーで出す。機能一覧 7）。付けた順。退会した利用者は含めない
            （そのため count より少ないことがある）
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
...
    MessageMention:
      description: メンションの対象（表示時の参照先。機能一覧 9.1 の経路2。対象がいまそのチャンネルの参加者かは問わない）
      type: object
      additionalProperties: false
      required:
        - userId
        - user
      properties:
        userId:
          description: 対象のユーザーID（登録したときの大文字小文字のまま）。本文の `@ユーザーID` とは大文字小文字によらず照合する（機能一覧 1.1）
          type: string
        user:
          description: 対象の利用者。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
...
    PinnedMessage:
      description: ピン留めしたメッセージ（機能一覧 13.2）
      type: object
      additionalProperties: false
      required:
        - message
        - pinnedBy
        - pinnedAt
      properties:
        message:
          $ref: '#/components/schemas/Message'
        pinnedBy:
          description: ピン留めした人。退会した利用者なら null（削除済みの利用者として表示する。機能一覧 1.5）
          oneOf:
            - $ref: '#/components/schemas/UserSummary'
            - type: 'null'
        pinnedAt:
          description: ピン留めした時刻
          type: string
          format: date-time
    PinList:
      type: object
      additionalProperties: false
      required:
        - pins
      properties:
        pins:
          description: ピン留めした新しい順
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/PinnedMessage'
...